Moved Annotation document to a different package with its subdocuments.
