+++ src/main/java/uk/ac/ebi/eva/pipeline/configuration/writers/AnnotationCompositeWriterConfiguration.java
@@ -25,7 +25,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Profile;
 
-import uk.ac.ebi.eva.commons.models.data.Annotation;
+import uk.ac.ebi.eva.commons.models.mongo.documents.Annotation;
 import uk.ac.ebi.eva.pipeline.Application;
 
 import java.util.Arrays;
