+++ src/main/java/uk/ac/ebi/eva/pipeline/configuration/writers/AnnotationInVariantWriterConfiguration.java
@@ -22,7 +22,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.mongodb.core.MongoOperations;
 
-import uk.ac.ebi.eva.commons.models.data.Annotation;
+import uk.ac.ebi.eva.commons.models.mongo.documents.Annotation;
 import uk.ac.ebi.eva.pipeline.Application;
 import uk.ac.ebi.eva.pipeline.io.writers.AnnotationInVariantMongoWriter;
 import uk.ac.ebi.eva.pipeline.parameters.AnnotationParameters;
