+++ src/main/java/uk/ac/ebi/eva/pipeline/io/writers/AnnotationInVariantMongoWriter.java
@@ -27,12 +27,12 @@ import org.springframework.data.mongodb.core.query.BasicUpdate;
 import org.springframework.util.Assert;
 
 import uk.ac.ebi.eva.commons.models.converters.data.VariantToDBObjectConverter;
-import uk.ac.ebi.eva.commons.models.data.Annotation;
+import uk.ac.ebi.eva.commons.models.mongo.documents.Annotation;
 import uk.ac.ebi.eva.commons.models.data.AnnotationFieldNames;
-import uk.ac.ebi.eva.commons.models.data.ConsequenceType;
+import uk.ac.ebi.eva.commons.models.mongo.documents.subdocuments.ConsequenceType;
 import uk.ac.ebi.eva.commons.models.data.Score;
 import uk.ac.ebi.eva.commons.models.data.VariantAnnotation;
-import uk.ac.ebi.eva.commons.models.data.Xref;
+import uk.ac.ebi.eva.commons.models.mongo.documents.subdocuments.Xref;
 import uk.ac.ebi.eva.utils.MongoDBHelper;
 
 import java.util.ArrayList;
